requirements.txt
alabaster==0.7.12
amqp==2.6.1
attrs==20.3.0
Babel==2.9.0
billiard==3.6.3.0
celery==4.4.7
certifi==2020.11.8
chardet==3.0.4
decorator==4.4.2
Django==2.2.14
django-activity-stream==0.9.0
django-celery-beat==2.0.0
django-docs==0.3.1
django-extensions==3.1.0
django-filebrowser==3.13.1
django-filter==2.3.0
django-grappelli==2.14.2
django-js-asset==1.2.2
django-mdeditor==0.1.18
django-model-utils==4.0.0
django-mptt==0.11.0
django-redis==4.12.1
django-taggit==1.3.0
django-taggit-serializer==0.1.7
django-timezone-field==4.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-recursive==0.1.2
docutils==0.16
drf-extensions==0.6.0
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
feedparser==5.2.1
flower==0.9.5
gunicorn==20.0.4
humanize==3.1.0
idna==2.10
imagesize==1.2.0
importlib-metadata==1.7.0
iniconfig==1.1.1
Jinja2==2.11.2
jsonpath==0.82
jsonpath-rw==1.4.0
kombu==4.6.11
livereload==2.6.3
Markdown==3.2.2
MarkupSafe==1.1.1
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
ply==3.11
prometheus-client==0.8.0
psycopg2-binary==2.8.5
py==1.9.0
Pygments==2.7.3
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.4.3
sphinx-autobuild==2020.9.1
sphinx-rtd-theme==0.5.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
toml==0.10.2
tornado==6.1
urllib3==1.25.11
uWSGI==2.0.19.1
vine==1.3.0
wrapt==1.12.1
zipp==3.1.0
